Scope web-mode display scale away from Tauri, split Tauri chrome into components
.shell's transform:scale() (web-mode display scale) is now scoped to
:global(html:not(.tauri)) — Tauri never gets it, so it can't hit the
containing-block side effect that misplaced the CSD resize grips and
would have similarly affected every other position:fixed overlay nested
in .shell (context menu, settings panel, rename dialog, status popover).
Tauri's .shell is back to its original pre-scale sizing.

Extracts the Tauri-only pieces into their own files instead of more
isTauri branches scattered through one 4000+ line component:
- lib/platform.ts: isTauri, scale application, settings load/save,
  shellOpen, winDrag — one place for the Tauri-vs-web branching.
- lib/components/TauriResizeGrips.svelte: the 8 CSD resize handles,
  rendered as a sibling of .shell rather than nested inside it.
- lib/components/TauriWindowControls.svelte: minimize/maximize/close.

Also fixes the status dot sitting flush against the browser's right edge
in web mode — .titlebar had no right padding, and Tauri's window-controls
buttons were the only thing providing a right-side buffer after it.

Fixes .gitignore's bare lib/ and lib64/ entries (Python build-artifact
boilerplate) silently ignoring ui/src/lib/, SvelteKit's $lib directory —
scoped to /lib/ and /lib64/ (repo-root only).
